requirements.txt
Pillow==8.3.0
